refactor: simplifying device into just device
